requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pandas==2.1.3
python-multipart==0.0.6
pytz==2023.3
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
alembic==1.12.1
psycopg2-binary==2.9.9 